Move item details list header and quantity control components to separate files. Add proptypes to components
